requirements.txt
fastapi
uvicorn
openai
requests
telethon
schedule
Pillow
pyrogram
Flask
Flask-SQLAlchemy